+++ app/src/main/java/bitcamp/project1/vo/Entry.java
@@ -3,10 +3,10 @@ package bitcamp.project1.vo;
 public class Entry {
   String type; // "수입" 또는 "지출"
   String description;
-  int amount;
+  double amount;
   String date;
 
-  public Entry(String type, String description, int amount, String date) {
+  public Entry(String type, String description, double amount, String date) {
     this.type = type;
     this.description = description;
     this.amount = amount;
